update importer and rename some stuff
