+++ scripts/buildEntityMap.js
@@ -1,10 +1,13 @@
-const fs = require('fs')
-fs.readFile('data/tracker_lists/disconnect-entitylist.json', 'utf8', function (err,data) {
+var request = require('request'),
+    fs = require('fs');
+
+request('https://raw.githubusercontent.com/mozilla-services/shavar-prod-lists/master/disconnect-entitylist.json', function (err, res, body) {
+
       if (err) {
           return console.log(err);
       }
 
-      let json = JSON.parse(data);
+      let json = JSON.parse(body);
       let out = {};
 
       for(let parent in json) {
